chore(week-5): formatted contracts

+++ src/week_5/contracts/Arrays.sol
@@ -2,11 +2,11 @@
 pragma solidity ^0.8.4;
 
 contract FixedSum {
-    function sum(uint[5] memory numbers) external pure returns(uint){
-        uint res = 0;
+    function sum(uint256[5] memory numbers) external pure returns (uint256) {
+        uint256 res = 0;
 
-        for (uint i = 0; i<numbers.length;i++){
-            res+=numbers[i];
+        for (uint256 i = 0; i < numbers.length; i++) {
+            res += numbers[i];
         }
 
         return res;
@@ -14,12 +14,10 @@ contract FixedSum {
 }
 
 contract DynamicSum {
-    
-
-    function sum(uint[] memory numbers) external pure returns(uint){
-        uint res = 0;
-        for(uint i =0;i< numbers.length;i++){
-            res+=numbers[i];
+    function sum(uint256[] memory numbers) external pure returns (uint256) {
+        uint256 res = 0;
+        for (uint256 i = 0; i < numbers.length; i++) {
+            res += numbers[i];
         }
 
         return res;
@@ -27,40 +25,38 @@ contract DynamicSum {
 }
 
 contract FilterToStorage {
-    uint[] public evenNumbers;
+    uint256[] public evenNumbers;
 
-    function filterEven(uint[] memory numbers) external{
-
-        for(uint i =0; i< numbers.length;i++){
-            if(numbers[i]%2 ==0){
+    function filterEven(uint256[] memory numbers) external {
+        for (uint256 i = 0; i < numbers.length; i++) {
+            if (numbers[i] % 2 == 0) {
                 evenNumbers.push(numbers[i]);
             }
         }
     }
 }
 
-
 contract FilterToMemory {
-    function filterEven(uint[] memory numbers) external pure returns(uint[] memory){
-         uint arrSize = 0;
+    function filterEven(uint256[] memory numbers) external pure returns (uint256[] memory) {
+        uint256 arrSize = 0;
 
-         for(uint i =0; i<numbers.length;i++){
-             if(numbers[i]%2==0){
-                 arrSize++;
-             }
-         }
+        for (uint256 i = 0; i < numbers.length; i++) {
+            if (numbers[i] % 2 == 0) {
+                arrSize++;
+            }
+        }
 
-        uint[] memory res = new uint[](arrSize);
+        uint256[] memory res = new uint[](arrSize);
 
-        uint currIdx = 0;
-        for(uint i =0; i<numbers.length;i++){
-             if(numbers[i]%2==0){
+        uint256 currIdx = 0;
+        for (uint256 i = 0; i < numbers.length; i++) {
+            if (numbers[i] % 2 == 0) {
                 res[currIdx] = numbers[i];
                 currIdx++;
-             }
-         }
+            }
+        }
 
-         return res;
+        return res;
     }
 }
 
@@ -68,10 +64,9 @@ contract StackClub {
     address[] public members;
     mapping(address => bool) public isMember;
 
-    constructor(){
+    constructor() {
         members.push(msg.sender);
         isMember[msg.sender] = true;
-
     }
 
     function addMember(address _newMember) external {
@@ -83,9 +78,9 @@ contract StackClub {
 
     function removeLastMember() external {
         require(isMember[msg.sender]);
-        require(members.length>1);
-        
-        address noMember = members[members.length-1];
+        require(members.length > 1);
+
+        address noMember = members[members.length - 1];
         members.pop();
         delete isMember[noMember];
     }
